Cuando hemos aplicado un filtro a un etiquetado de una capa ya no es posible quitarlo, hay que crear simbología nueva. Refs:#8614 - #8837
git-svn-id: https://devel.scolab.eu/gvsig-online/gvsigonline_v2/trunk@7988 216424d3-d1a8-4ca4-951f-a609719bba38

+++ gvsigol/gvsigol_symbology/services.py
@@ -545,29 +545,30 @@ def clone_style(mapservice, layer, original_style_name, cloned_style_name):
     return True
 
      
-#eliminar estilo solo de gvsigonline    
+#eliminar estilo solo de gvsigonline, no llego a guardarlo en geoserver    
 def delete_style_name(name):
     try:
         last_inserted_id = Style.objects.filter(name=name).aggregate(Max('id'))['id__max']
 
-        style = Style.objects.get(id=int(last_inserted_id))
+        if last_inserted_id is not None:
+           style = Style.objects.get(id=int(last_inserted_id))
         
-      
-        layer_styles = StyleLayer.objects.filter(style=style)   
-        for layer_style in layer_styles:
-            layer_style.delete()
-                
-        rules = Rule.objects.filter(style=style)
-        for rule in rules:
-            symbolizers = Symbolizer.objects.filter(rule=rule)
-            for symbolizer in symbolizers:
-                if hasattr(symbolizer, 'rastersymbolizer'):
-                    symbolizer.rastersymbolizer.color_map.delete()
-                symbolizer.delete()
-            rule.delete()
-    
-        style.delete()
-        
+           if style:
+                layer_styles = StyleLayer.objects.filter(style=style)   
+                for layer_style in layer_styles:
+                    layer_style.delete()
+                        
+                rules = Rule.objects.filter(style=style)
+                for rule in rules:
+                    symbolizers = Symbolizer.objects.filter(rule=rule)
+                    for symbolizer in symbolizers:
+                        if hasattr(symbolizer, 'rastersymbolizer'):
+                            symbolizer.rastersymbolizer.color_map.delete()
+                        symbolizer.delete()
+                    rule.delete()
+            
+                style.delete()
+            
     except Exception as e:
         raise e
